Break v2 API out into separate module
