+++ conda-store-server/conda_store_server/_internal/server/views/__init__.py
@@ -2,7 +2,8 @@
 # Use of this source code is governed by a BSD-style
 # license that can be found in the LICENSE file.
 
-from conda_store_server._internal.server.views.api import router_apis  # noqa
+from conda_store_server._internal.server.views.api import router_api as router_api_v1  # noqa
+from conda_store_server._internal.server.views.api_v2 import router_api as router_api_v2  # noqa
 from conda_store_server._internal.server.views.conda_store_ui import (  # noqa
     router_conda_store_ui,
 )
